+++ packages/compiler-cli/ngcc/src/dependencies/module_resolver.ts
@@ -6,7 +6,8 @@
  * found in the LICENSE file at https://angular.io/license
  */
 import {absoluteFrom, AbsoluteFsPath, dirname, FileSystem, isRoot, join, resolve} from '../../../src/ngtsc/file_system';
-import {isRelativePath, PathMappings, resolveFileWithPostfixes} from '../utils';
+import {PathMappings} from '../ngcc_options';
+import {isRelativePath, resolveFileWithPostfixes} from '../utils';
 
 /**
  * This is a very cut-down implementation of the TypeScript module resolution strategy.
